cambiato tipo di dato per telefono in database

+++ database/e-spresso.sql
@@ -21,7 +21,7 @@ CREATE TABLE users (
 	indirizzo VARCHAR(100) NOT NULL,
 	n_civico VARCHAR(5) NOT NULL,
 	citta VARCHAR(100) NOT NULL,
-	numero_telefono INT NOT NULL
+	numero_telefono NUMERIC(10) NOT NULL
 );
 
 -- TABELLA CARRELLO:
